plugin tests passing; use `ACTIONHERO_CONFIG_OVERRIDES` to change config
